test: fix a few mutants from `cargo-mutants`
Using `cargo mutants -p firezone-gui-client -p firezone-headless-client`

155 mutants missed after
