+++ rust/headless-client/src/known_dirs.rs
@@ -10,131 +10,16 @@
 pub use platform::{ipc_service_logs, logs, runtime, session, settings};
 
 #[cfg(target_os = "linux")]
-pub mod platform {
-    use connlib_shared::BUNDLE_ID;
-    use std::path::PathBuf;
-
-    #[allow(clippy::unnecessary_wraps)]
-    pub fn ipc_service_logs() -> Option<PathBuf> {
-        // TODO: This is magic, it must match the systemd file
-        Some(PathBuf::from("/var/log").join(connlib_shared::BUNDLE_ID))
-    }
-
-    /// e.g. `/home/alice/.cache/dev.firezone.client/data/logs`
-    ///
-    /// Logs are considered cache because they're not configs and it's technically okay
-    /// if the system / user deletes them to free up space
-    pub fn logs() -> Option<PathBuf> {
-        Some(dirs::cache_dir()?.join(BUNDLE_ID).join("data").join("logs"))
-    }
-
-    /// e.g. `/run/user/1000/dev.firezone.client/data`
-    ///
-    /// Crash handler socket and other temp files go here
-    pub fn runtime() -> Option<PathBuf> {
-        Some(dirs::runtime_dir()?.join(BUNDLE_ID).join("data"))
-    }
-
-    /// e.g. `/home/alice/.local/share/dev.firezone.client/data`
-    ///
-    /// Things like actor name are stored here because they're kind of config,
-    /// the system / user should not delete them to free up space, but they're not
-    /// really config since the program will rewrite them automatically to persist sessions.
-    pub fn session() -> Option<PathBuf> {
-        Some(dirs::data_local_dir()?.join(BUNDLE_ID).join("data"))
-    }
-
-    /// e.g. `/home/alice/.config/dev.firezone.client/config`
-    ///
-    /// See connlib docs for details
-    pub fn settings() -> Option<PathBuf> {
-        Some(dirs::config_local_dir()?.join(BUNDLE_ID).join("config"))
-    }
-}
+#[path = "known_dirs/linux.rs"]
+pub mod platform;
 
 #[cfg(target_os = "macos")]
-pub mod platform {
-    pub fn ipc_service_logs() -> Option<PathBuf> {
-        unimplemented!()
-    }
-
-    pub fn logs() -> Option<PathBuf> {
-        unimplemented!()
-    }
-
-    pub fn runtime() -> Option<PathBuf> {
-        unimplemented!()
-    }
-
-    pub fn session() -> Option<PathBuf> {
-        unimplemented!()
-    }
-
-    pub fn settings() -> Option<PathBuf> {
-        unimplemented!()
-    }
-}
+#[path = "known_dirs/macos.rs"]
+pub mod platform;
 
 #[cfg(target_os = "windows")]
-pub mod platform {
-    use connlib_shared::BUNDLE_ID;
-    use known_folders::{get_known_folder_path, KnownFolder};
-    use std::path::PathBuf;
-
-    pub fn ipc_service_logs() -> Option<PathBuf> {
-        Some(
-            get_known_folder_path(KnownFolder::ProgramData)?
-                .join(BUNDLE_ID)
-                .join("data")
-                .join("logs"),
-        )
-    }
-
-    /// e.g. `C:\Users\Alice\AppData\Local\dev.firezone.client\data\logs`
-    ///
-    /// See connlib docs for details
-    pub fn logs() -> Option<PathBuf> {
-        Some(
-            connlib_shared::windows::app_local_data_dir()
-                .ok()?
-                .join("data")
-                .join("logs"),
-        )
-    }
-
-    /// e.g. `C:\Users\Alice\AppData\Local\dev.firezone.client\data`
-    ///
-    /// Crash handler socket and other temp files go here
-    pub fn runtime() -> Option<PathBuf> {
-        Some(
-            connlib_shared::windows::app_local_data_dir()
-                .ok()?
-                .join("data"),
-        )
-    }
-
-    /// e.g. `C:\Users\Alice\AppData\Local\dev.firezone.client\data`
-    ///
-    /// Things like actor name go here
-    pub fn session() -> Option<PathBuf> {
-        Some(
-            connlib_shared::windows::app_local_data_dir()
-                .ok()?
-                .join("data"),
-        )
-    }
-
-    /// e.g. `C:\Users\Alice\AppData\Local\dev.firezone.client\config`
-    ///
-    /// See connlib docs for details
-    pub fn settings() -> Option<PathBuf> {
-        Some(
-            connlib_shared::windows::app_local_data_dir()
-                .ok()?
-                .join("config"),
-        )
-    }
-}
+#[path = "known_dirs/windows.rs"]
+pub mod platform;
 
 #[cfg(test)]
 mod tests {
