+++ rust/headless-client/src/known_dirs/macos.rs
@@ -1,0 +1,19 @@
+pub fn ipc_service_logs() -> Option<PathBuf> {
+    unimplemented!()
+}
+
+pub fn logs() -> Option<PathBuf> {
+    unimplemented!()
+}
+
+pub fn runtime() -> Option<PathBuf> {
+    unimplemented!()
+}
+
+pub fn session() -> Option<PathBuf> {
+    unimplemented!()
+}
+
+pub fn settings() -> Option<PathBuf> {
+    unimplemented!()
+}
