+++ rust/headless-client/src/windows.rs
@@ -299,5 +299,7 @@ mod tests {
             super::named_pipe_path("dev.firezone.client"),
             r"\\.\pipe\dev.firezone.client"
         );
+
+        assert!(super::pipe_path().starts_with(r"\\.\pipe\"));
     }
 }
